# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "imperia-api",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Mapped server"
        },
        {
            "url": "http://localhost:8080",
            "description": "Current server"
        },
        {
            "url": "https://imperia-api.herokuapp.com",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Index services.",
                "operationId": "indexServices",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ServiceIncludes"
                        }
                    },
                    {
                        "name": "page[size]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageSize"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageNumber"
                        }
                    },
                    {
                        "name": "filter[title]",
                        "in": "query",
                        "description": "Can be used for searches. Is partial.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Clown"
                    },
                    {
                        "name": "filter[categories]",
                        "in": "query",
                        "description": "Coma-separated array of category ids. Limits services to those\n     * that have at least one of given categories attached to them",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2,3"
                    },
                    {
                        "name": "archived",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ArchivedParameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index services response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexServiceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/services/{id}": {
            "get": {
                "tags": [
                    "services"
                ],
                "summary": "Show service by id.",
                "operationId": "showService",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the services.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ServiceIncludes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show services response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowServiceResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "Index tickets.",
                "operationId": "indexTickets",
                "parameters": [
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/TicketIncludes"
                        }
                    },
                    {
                        "name": "page[size]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageSize"
                        }
                    },
                    {
                        "name": "page[number]",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/PageNumber"
                        }
                    },
                    {
                        "name": "filter[title]",
                        "in": "query",
                        "description": "Can be used for searches. Is partial.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Weekend"
                    },
                    {
                        "name": "filter[categories]",
                        "in": "query",
                        "description": "Coma-separated array of category ids. Limits tickets to those\n     * that have at least one of given categories attached to them",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "2,3"
                    },
                    {
                        "name": "archived",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/ArchivedParameter"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Index tickets response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IndexTicketResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "tickets"
                ],
                "summary": "Show ticket by id.",
                "operationId": "showTicket",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the tickets.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "$ref": "#/components/schemas/TicketIncludes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show tickets response object.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ShowTicketResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "description": "Unauthenticated response object.",
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    }
                },
                "type": "object"
            },
...
            "IndexServiceResponse": {
                "description": "Index services response object.",
                "required": [
                    "data",
                    "meta",
                    "message"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Service"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ShowServiceResponse": {
                "description": "Show services response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Service"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ServiceIncludes": {
                "description": "Coma-separated list of inluded relations.\n    Available relations: `categories`",
                "type": "string",
                "example": "categories"
            },
...
            "IndexTicketResponse": {
                "description": "Index tickets response object.",
                "required": [
                    "data",
                    "meta",
                    "message"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Ticket"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "ShowTicketResponse": {
                "description": "Show tickets response object.",
                "required": [
                    "data",
                    "message"
                ],
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Ticket"
                    },
                    "message": {
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "TicketIncludes": {
                "description": "Coma-separated list of inluded relations.\n    Available relations: `categories`",
                "type": "string",
                "example": "categories"
            },
...
            "PageSize": {
                "description": "Max number of results on page.",
                "type": "integer",
                "example": 25
            },
            "PageNumber": {
                "description": "Page number to be returned.",
                "type": "integer",
                "example": 1
            },
...
            "Category": {
                "description": "Category resource object",
                "required": [
                    "id",
                    "type",
                    "slug",
                    "target",
                    "title",
                    "description",
                    "media",
                    "default_media"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "pizza"
                    },
                    "type": {
                        "type": "string",
                        "example": "categories"
                    },
                    "target": {
                        "nullable": "true",
                        "type": "string",
                        "example": "products"
                    },
                    "title": {
                        "type": "string",
                        "example": "Pizza"
                    },
                    "description": {
                        "nullable": "true",
                        "example": "Some text..."
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    },
                    "default_media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    }
                },
                "type": "object"
            },
...
            "Media": {
                "description": "Media resource object",
                "required": [
                    "id",
                    "type",
                    "name",
                    "title",
                    "url",
                    "path",
                    "folder",
                    "private"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "media"
                    },
                    "name": {
                        "type": "string",
                        "example": "mojito-1645908379CXJyz.jpg"
                    },
                    "title": {
                        "type": "string",
                        "example": "Mojito"
                    },
                    "url": {
                        "type": "string",
                        "example": "http://localhost/storage/mojito-1645908379CXJyz.jpg"
                    },
                    "path": {
                        "type": "string",
                        "example": "/mojito-1645908379CXJyz.jpg"
                    },
                    "folder": {
                        "type": "string",
                        "example": "/"
                    },
                    "private": {
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            },
...
            "Links": {
                "description": "Pagination meta links object.",
                "required": [
                    "first",
                    "prev",
                    "self",
                    "next",
                    "last"
                ],
                "properties": {
                    "first": {
                        "type": "string"
                    },
                    "prev": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "self": {
                        "type": "string"
                    },
                    "next": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    },
                    "last": {
                        "nullable": "true",
                        "type": "string",
                        "example": null
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "description": "Pagination meta object.",
                "required": [
                    "from",
                    "to",
                    "total",
                    "path",
                    "per_page",
                    "current_page",
                    "last_page",
                    "links"
                ],
                "properties": {
                    "from": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "to": {
                        "nullable": "true",
                        "type": "integer",
                        "example": 1
                    },
                    "total": {
                        "type": "integer",
                        "example": 30
                    },
                    "path": {
                        "nullable": "true",
                        "type": "integer",
                        "example": null
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 2
                    },
                    "links": {
                        "$ref": "#/components/schemas/Links"
                    }
                },
                "type": "object"
            },
            "Service": {
                "description": "Service resource object",
                "required": [
                    "id",
                    "type",
                    "title",
                    "description",
                    "once_paid_price",
                    "hourly_paid_price",
                    "category_ids",
                    "media",
                    "default_media"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "services"
                    },
                    "title": {
                        "type": "string",
                        "example": "Clown Show"
                    },
                    "description": {
                        "type": "string",
                        "example": "Some text..."
                    },
                    "once_paid_price": {
                        "type": "float",
                        "example": 350
                    },
                    "hourly_paid_price": {
                        "type": "float",
                        "example": 200
                    },
                    "archived": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    },
                    "default_media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    }
                },
                "type": "object"
            },
...
            "Ticket": {
                "description": "Ticket resource object",
                "required": [
                    "id",
                    "type",
                    "title",
                    "description",
                    "price",
                    "category_ids",
                    "media",
                    "default_media"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "type": {
                        "type": "string",
                        "example": "tickets"
                    },
                    "title": {
                        "type": "string",
                        "example": "Child ticket"
                    },
                    "description": {
                        "type": "string",
                        "example": "Some text..."
                    },
                    "price": {
                        "type": "float",
                        "example": 19.6
                    },
                    "archived": {
                        "type": "boolean",
                        "example": "false"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Category"
                        }
                    },
                    "category_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    },
                    "default_media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        }
                    }
                },
                "type": "object"
            },
...
            "ArchivedParameter": {
                "description": "Query parameter, which determines if archived records should be\n    fetched from database. Available relations: `only`, `with` and `without`, which is a default one.",
                "type": "string",
                "enum": [
                    "only",
                    "with",
                    "without"
                ],
                "example": "without"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}